+++ sbin/fsck_ffs/main.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: main.c,v 1.38 2011/04/24 07:07:03 otto Exp $	*/
+/*	$OpenBSD: main.c,v 1.39 2013/04/25 06:43:20 otto Exp $	*/
 /*	$NetBSD: main.c,v 1.22 1996/10/11 20:15:48 thorpej Exp $	*/
 
 /*
@@ -262,7 +262,8 @@ checkfilesys(char *filesys, char *mntpt, long auxdata, int child)
 		if (zlnhead != NULL) {
 			printf("The following zero link count inodes remain:");
 			for (zlnp = zlnhead; zlnp; zlnp = zlnp->next)
-				printf(" %u,", zlnp->zlncnt);
+				printf(" %llu,",
+				    (unsigned long long)zlnp->zlncnt);
 			printf("\n");
 		}
 	}
